+++ crates/canic/src/lib.rs
@@ -42,7 +42,9 @@ pub mod prelude {
             access::auth::{is_controller, is_parent, is_root},
             auth_require_all, auth_require_any,
             ids::CanisterRole,
-            log, perf, perf_scope, timer, timer_interval,
+            log,
+            ops::ic::call::Call,
+            perf, perf_scope, timer, timer_interval,
         },
         macros::{canic_query, canic_update},
     };